Massive refactor. - Use QDrant - Use llama-cpp
So how does this work?

Outside of this container (in the qdrant container) there now is a database with all the IMDB series and movie titles.
It is stored as embeddings. This lookup mechanish (as before) cleans the input to detect (on a best effort basis):
- year
- title
- series or movie
It then creates an embedding out of that (llama-cpp) which is send to QDrant to give the best matches.
The return from qdrant is then again matched by similarity compared to your input title.

This results in an IMDB ID. If the lookup is just for an IMDB ID (so a POST to the / endpoint) then this lookup is now completely free of external services.
The QDrant database is internally hosted. When more data from IMDB is needed (POST to /full or any of the GET lookups) then there is potentially a lookup to The Movie DB. Unless it's cached then that too is free.

In all, the lookup accuracy is now insanely high.