Even more substantially improved!
This time the improvements are about the limit of what's possible given the dataset.
Anything that should get an error (as in can't determine IMDB ID from download name) will likely get that error.
And false positives are likely real false possitives unable to solve. Think of two different series (or movies) with the exact same name. That will just be impossible to correctly determine.

Lastly the output of the episode specific data has been enriched a little. It now includes the series backdrop image, the series poster and the episode airtime.